feat: remove the builtin_ev package and move everything in it to handlers.
Also moves some small things in event, although not as relevant.
