--- internal/builtin/events/example
@@ -1,22 +1,0 @@
-package builtin_ev
-
-import (
-	"github.com/Nykenik24/enkrypted/internal/event"
-	"github.com/Nykenik24/enkrypted/internal/user"
-)
-
-var ExampleEventKind = buildKind(SomeNamespace, "example")
-
-type ExampleEvent struct{}
-
-func NewExampleEvent() *ExampleEvent {
-	return &ExampleEvent{}
-}
-
-func (ev *ExampleEvent) Data() *event.EventData {
-	return &event.EventData{}
-}
-
-func (ev *ExampleEvent) Kind() *event.EventKind {
-	return ExampleEventKind
-}
