--- internal/builtin/events/message.go
@@ -1,38 +1,0 @@
-package builtin_ev
-
-import (
-	"github.com/Nykenik24/enkrypted/internal/event"
-	"github.com/Nykenik24/enkrypted/internal/user"
-)
-
-var MessageEventKind = buildKind(RoomNamespace, "message")
-
-type MessageEvent struct {
-	Contents  string     `json:"contents"`
-	Timestamp string     `json:"timestamp"`
-	User      *user.User `json:"user"`
-	ID        uint64     `json:"id"`
-}
-
-func NewMessageEvent(contents, timestamp string, user *user.User) *MessageEvent {
-	lastMessageID++
-	return &MessageEvent{
-		Contents:  contents,
-		Timestamp: timestamp,
-		User:      user,
-		ID:        lastMessageID,
-	}
-}
-
-func (ev *MessageEvent) Data() *event.EventData {
-	return &event.EventData{
-		"contents":  ev.Contents,
-		"timestamp": ev.Timestamp,
-		"user":      ev.User,
-		"id":        ev.ID,
-	}
-}
-
-func (ev *MessageEvent) Kind() *event.EventKind {
-	return MessageEventKind
-}
