+++ internal/handlers/connect.go
@@ -4,22 +4,38 @@ import (
 	"fmt"
 	"time"
 
-	builtin_ev "github.com/Nykenik24/enkrypted/internal/builtin/events"
+	"github.com/Nykenik24/enkrypted/internal/event"
 	"github.com/Nykenik24/enkrypted/internal/ws"
 )
+
+var ConnectEventKind = buildKind(WebsocketNamespace, "connect")
+
+type ConnectEvent struct {
+	Timestamp string `json:"timestamp"`
+	ID        uint64 `json:"userId"`
+}
+
+func NewConnectEvent(id uint64) *ConnectEvent {
+	return &ConnectEvent{
+		Timestamp: time.Now().Format(time.RFC3339),
+		ID:        id,
+	}
+}
+
+func (ev *ConnectEvent) Data() *event.EventData {
+	return &event.EventData{
+		"timestamp": ev.Timestamp,
+		"userId":    ev.ID,
+	}
+}
+
+func (ev *ConnectEvent) Kind() *event.EventKind {
+	return ConnectEventKind
+}
 
 type ConnectHandler struct{}
 
 func (h *ConnectHandler) Handle(ctx *ws.Context) error {
-	ev := builtin_ev.NewMessageEvent(
-		fmt.Sprintf("user joined %s", ctx.Client.GetUser().Username),
-		time.Now().Format(time.RFC3339),
-		ctx.Client.GetUser(),
-	)
-
-	ctx.Broadcast(
-		builtin_ev.Generic(ev).ToAll(),
-	)
-
+	BroadcastMessage(ctx, fmt.Sprintf("user joined %s", ctx.Client.GetUser().Username))
 	return nil
 }
