--- internal/handlers/example
@@ -1,18 +1,0 @@
-package handlers
-
-import (
-	builtin_ev "github.com/Nykenik24/enkrypted/internal/builtin/events"
-	"github.com/Nykenik24/enkrypted/internal/ws"
-)
-
-type ExampleHandler struct{}
-
-func (h *ExampleHandler) Handle(ctx *ws.Context) error {
-	var ev builtin_ev.SomeEvent
-
-	if err := ctx.Bind(&ev); err != nil {
-		return err
-	}
-
-	return nil
-}
