+++ internal/handlers/shared.go
@@ -1,4 +1,4 @@
-package builtin_ev
+package handlers
 
 import (
 	"fmt"
@@ -32,7 +32,7 @@ type BuiltinEvent interface {
 	Kind() *event.EventKind
 }
 
-func Generic(ev BuiltinEvent) *event.Event {
+func Base(ev BuiltinEvent) *event.Event {
 	return &event.Event{
 		Kind: ev.Kind(),
 		Data: ev.Data(),
